NEW: Coherence estimate from single trial MTM cross-spectra
- new CR's and CF's for post-processing the parallelised single trial routines in 'AV_compRoutines'. They are intended to get called directly within the frontend on the ST_compRoutines results to perform computations after trial averaging. Right now only the `Norm_CrossMeasures` is implemented which computes coherences/correlations from the respective matrix quantities. Their input is intended to be strictly 1-Trial (after averaging) and don't trivially support parallelisation over channels, hence parallel=False is hardcoded for this post-processing methods.
- Added backend test for the new trial averaged coherence
- converted all remaining `np.complex128` to `np.complex64` in specest for consistency

Changes to be committed:
	modified:   ../../dev_frontend.py
	new file:   AV_compRoutines.py
	modified:   ST_compRoutines.py
	modified:   connectivity_analysis.py
	modified:   const_def.py
	modified:   ../specest/const_def.py
	modified:   ../specest/freqanalysis.py
	modified:   ../specest/mtmfft.py
	modified:   ../tests/backend/test_connectivity.py

+++ syncopy/specest/mtmfft.py
@@ -80,7 +80,7 @@ def mtmfft(data_arr, samplerate, taper="hann", taperopt={}):
         windows = windows * np.sqrt(nSamples) / np.sum(windows)
         
     # Fourier transforms (nTapers x nFreq x nChannels)
-    ftr = np.zeros((windows.shape[0], nFreq, nChannels), dtype='complex128')
+    ftr = np.zeros((windows.shape[0], nFreq, nChannels), dtype='complex64')
 
     for taperIdx, win in enumerate(windows):
         win = np.tile(win, (nChannels, 1)).T
